Move autocomplete to seperate file
